Adjust namespaces and extensions file splitting
