--- dotnet/src/Connectors/Connectors.AzureOpenAI/Extensions/AzureOpenAIKernelBuilderExtensions.EmbeddingGenerator.cs
@@ -1,127 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-using System.Diagnostics.CodeAnalysis;
-using System.Net.Http;
-using Azure.AI.OpenAI;
-using Azure.Core;
-using Microsoft.SemanticKernel.Connectors.AzureOpenAI;
-
-namespace Microsoft.SemanticKernel;
-
-/// <summary>
-/// Provides extension methods for <see cref="IKernelBuilder"/> to configure Azure OpenAI connectors.
-/// </summary>
-public static partial class AzureOpenAIKernelBuilderExtensions
-{
-    /// <summary>
-    /// Adds the <see cref="AzureOpenAITextEmbeddingGenerationService"/> to the <see cref="IKernelBuilder.Services"/>.
-    /// </summary>
-    /// <param name="builder">The <see cref="IKernelBuilder"/> instance to augment.</param>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="endpoint">Azure OpenAI deployment URL, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="apiKey">Azure OpenAI API key, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="serviceId">A local identifier for the given AI service</param>
-    /// <param name="modelId">Model identifier, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="dimensions">The number of dimensions the resulting output embeddings should have. Only supported in "text-embedding-3" and later models.</param>
-    /// <param name="apiVersion">Optional Azure OpenAI API version, see available here <see cref="AzureOpenAIClientOptions.ServiceVersion"/></param>
-    /// <param name="httpClient">The HttpClient to use with this service.</param>
-    /// <returns>The same instance as <paramref name="builder"/>.</returns>
-    [Experimental("SKEXP0010")]
-    public static IKernelBuilder AddAzureOpenAIEmbeddingGenerator(
-        this IKernelBuilder builder,
-        string deploymentName,
-        string endpoint,
-        string apiKey,
-        string? serviceId = null,
-        string? modelId = null,
-        int? dimensions = null,
-        string? apiVersion = null,
-        HttpClient? httpClient = null)
-    {
-        Verify.NotNull(builder);
-
-        builder.Services.AddAzureOpenAIEmbeddingGenerator(
-            deploymentName,
-            endpoint,
-            apiKey,
-            serviceId,
-            modelId,
-            dimensions,
-            apiVersion,
-            httpClient);
-
-        return builder;
-    }
-
-    /// <summary>
-    /// Adds the <see cref="AzureOpenAITextEmbeddingGenerationService"/> to the <see cref="IKernelBuilder.Services"/>.
-    /// </summary>
-    /// <param name="builder">The <see cref="IKernelBuilder"/> instance to augment.</param>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="endpoint">Azure OpenAI deployment URL, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="credential">Token credentials, e.g. DefaultAzureCredential, ManagedIdentityCredential, EnvironmentCredential, etc.</param>
-    /// <param name="serviceId">A local identifier for the given AI service</param>
-    /// <param name="modelId">Model identifier, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="dimensions">The number of dimensions the resulting output embeddings should have. Only supported in "text-embedding-3" and later models.</param>
-    /// <param name="apiVersion">Optional Azure OpenAI API version, see available here <see cref="AzureOpenAIClientOptions.ServiceVersion"/></param>
-    /// <param name="httpClient">The HttpClient to use with this service.</param>
-    /// <returns>The same instance as <paramref name="builder"/>.</returns>
-    [Experimental("SKEXP0010")]
-    public static IKernelBuilder AddAzureOpenAIEmbeddingGenerator(
-        this IKernelBuilder builder,
-        string deploymentName,
-        string endpoint,
-        TokenCredential credential,
-        string? serviceId = null,
-        string? modelId = null,
-        int? dimensions = null,
-        string? apiVersion = null,
-        HttpClient? httpClient = null)
-    {
-        Verify.NotNull(builder);
-        Verify.NotNull(credential);
-
-        builder.Services.AddAzureOpenAIEmbeddingGenerator(
-            deploymentName,
-            endpoint,
-            credential,
-            serviceId,
-            modelId,
-            dimensions,
-            apiVersion,
-            httpClient);
-
-        return builder;
-    }
-
-    /// <summary>
-    /// Adds the <see cref="AzureOpenAITextEmbeddingGenerationService"/> to the <see cref="IKernelBuilder.Services"/>.
-    /// </summary>
-    /// <param name="builder">The <see cref="IKernelBuilder"/> instance to augment.</param>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="azureOpenAIClient"><see cref="AzureOpenAIClient"/> to use for the service. If null, one must be available in the service provider when this service is resolved.</param>
-    /// <param name="serviceId">A local identifier for the given AI service</param>
-    /// <param name="modelId">Model identifier, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="dimensions">The number of dimensions the resulting output embeddings should have. Only supported in "text-embedding-3" and later models.</param>
-    /// <returns>The same instance as <paramref name="builder"/>.</returns>
-    [Experimental("SKEXP0010")]
-    public static IKernelBuilder AddAzureOpenAIEmbeddingGenerator(
-        this IKernelBuilder builder,
-        string deploymentName,
-        AzureOpenAIClient? azureOpenAIClient = null,
-        string? serviceId = null,
-        string? modelId = null,
-        int? dimensions = null)
-    {
-        Verify.NotNull(builder);
-
-        builder.Services.AddAzureOpenAIEmbeddingGenerator(
-            deploymentName,
-            azureOpenAIClient,
-            serviceId,
-            modelId,
-            dimensions);
-
-        return builder;
-    }
-}
