+++ dotnet/src/IntegrationTests/Connectors/MistralAI/TextEmbedding/MistralAITextEmbeddingTests.cs
@@ -46,4 +46,23 @@ public sealed class MistralAITextEmbeddingGenerationServiceTests
         Assert.Equal(1024, response[0].Length);
         Assert.Equal(1024, response[1].Length);
     }
+
+    [Fact(Skip = "This test is for manual verification.")]
+    public async Task MistralAIEmbeddingGeneratorAsync()
+    {
+        // Arrange
+        var model = this._configuration["MistralAI:EmbeddingModel"];
+        var apiKey = this._configuration["MistralAI:ApiKey"];
+        using var service = new MistralAIEmbeddingGenerator(model!, apiKey!);
+
+        // Act
+        List<string> data = ["Hello", "world"];
+        var response = (await service.GenerateAsync(data));
+
+        // Assert
+        Assert.NotNull(response);
+        Assert.Equal(2, response.Count);
+        Assert.Equal(1024, response[0].Vector.Length);
+        Assert.Equal(1024, response[1].Vector.Length);
+    }
 }
